+++ src/main/java/com/fab/banggabgo/dto/sign/NickNameCheckResultDto.java
@@ -1,4 +1,4 @@
-package com.fab.banggabgo.dto;
+package com.fab.banggabgo.dto.sign;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class NameCheckResultDto {
+public class NickNameCheckResultDto {
 
   private String msg;
 }
