+++ src/main/java/com/fab/banggabgo/dto/sign/SignInRequestForm.java
@@ -1,36 +1,26 @@
-package com.fab.banggabgo.dto;
+package com.fab.banggabgo.dto.sign;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
-@Getter
-@Setter
 @Builder
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class SignUpRequestForm {
+public class SignInRequestForm {
 
   @ApiModelProperty(value = "이메일", example = "test@test.com")
-
   private String email;
   @ApiModelProperty(value = "비밀번호", example = "MyPassword@123")
-
   private String password;
-  @ApiModelProperty(value = "닉네임", example = "수줍은라이언")
 
-  private String nickname;
-
-  public static SignUpRequestDto toDto(SignUpRequestForm form) {
-    return SignUpRequestDto.builder()
-        .email(form.email)
-        .password(form.password)
-        .nickname(form.getNickname())
+  public static SignInRequestDto toDto(SignInRequestForm form) {
+    return SignInRequestDto.builder()
+        .email(form.getEmail())
+        .password(form.getPassword())
         .build();
   }
 }
-
-
