+++ src/main/java/com/fab/banggabgo/dto/sign/TokenDto.java
@@ -1,4 +1,4 @@
-package com.fab.banggabgo.dto;
+package com.fab.banggabgo.dto.sign;
 
 import lombok.Builder;
 import lombok.Getter;
