+++ src/main/java/com/fab/banggabgo/service/SignService.java
@@ -1,11 +1,11 @@
 package com.fab.banggabgo.service;
 
-import com.fab.banggabgo.dto.EmailCheckResultDto;
-import com.fab.banggabgo.dto.LogOutResultDto;
-import com.fab.banggabgo.dto.NameCheckResultDto;
-import com.fab.banggabgo.dto.SignInRequestDto;
-import com.fab.banggabgo.dto.SignInResultDto;
-import com.fab.banggabgo.dto.SignUpRequestDto;
+import com.fab.banggabgo.dto.sign.EmailCheckResultDto;
+import com.fab.banggabgo.dto.sign.LogOutResultDto;
+import com.fab.banggabgo.dto.sign.NickNameCheckResultDto;
+import com.fab.banggabgo.dto.sign.SignInRequestDto;
+import com.fab.banggabgo.dto.sign.SignInResultDto;
+import com.fab.banggabgo.dto.sign.SignUpRequestDto;
 import javax.servlet.http.HttpServletRequest;
 
 public interface SignService {
@@ -20,5 +20,5 @@ public interface SignService {
   EmailCheckResultDto emailCheck(String email);
 
 
-  NameCheckResultDto nickNameCheck(String name);
+  NickNameCheckResultDto nickNameCheck(String name);
 }
