+++ src/main/java/com/fab/banggabgo/service/impl/SignServiceImpl.java
@@ -3,13 +3,13 @@ package com.fab.banggabgo.service.impl;
 import com.fab.banggabgo.common.exception.CustomException;
 import com.fab.banggabgo.common.exception.ErrorCode;
 import com.fab.banggabgo.config.security.JwtTokenProvider;
-import com.fab.banggabgo.dto.EmailCheckResultDto;
-import com.fab.banggabgo.dto.LogOutResultDto;
-import com.fab.banggabgo.dto.NameCheckResultDto;
-import com.fab.banggabgo.dto.SignInRequestDto;
-import com.fab.banggabgo.dto.SignInResultDto;
-import com.fab.banggabgo.dto.SignUpRequestDto;
-import com.fab.banggabgo.dto.TokenDto;
+import com.fab.banggabgo.dto.sign.EmailCheckResultDto;
+import com.fab.banggabgo.dto.sign.LogOutResultDto;
+import com.fab.banggabgo.dto.sign.NickNameCheckResultDto;
+import com.fab.banggabgo.dto.sign.SignInRequestDto;
+import com.fab.banggabgo.dto.sign.SignInResultDto;
+import com.fab.banggabgo.dto.sign.SignUpRequestDto;
+import com.fab.banggabgo.dto.sign.TokenDto;
 import com.fab.banggabgo.entity.User;
 import com.fab.banggabgo.repository.UserRepository;
 import com.fab.banggabgo.service.SignService;
@@ -118,9 +118,9 @@ public class SignServiceImpl implements SignService {
   }
 
   @Override
-  public NameCheckResultDto nickNameCheck(String nickname) {
+  public NickNameCheckResultDto nickNameCheck(String nickname) {
     checkDuplicate(nickname, ErrorCode.NICKNAME_ALREADY_EXISTS);
-    return NameCheckResultDto.builder()
+    return NickNameCheckResultDto.builder()
         .msg("사용 가능한 별명 입니다.")
         .build();
   }
