+++ src/test/java/com/fab/banggabgo/controller/ArticleControllerTest.java
@@ -3,7 +3,6 @@ package com.fab.banggabgo.controller;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doThrow;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -14,8 +13,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fab.banggabgo.common.exception.CustomException;
 import com.fab.banggabgo.common.exception.ErrorCode;
-import com.fab.banggabgo.dto.ArticleEditForm;
-import com.fab.banggabgo.dto.ArticleRegisterForm;
+import com.fab.banggabgo.dto.article.ArticleEditForm;
+import com.fab.banggabgo.dto.article.ArticleRegisterForm;
 import com.fab.banggabgo.service.ArticleService;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
